grass.
Poaceae (/poʊˈeɪsiaɪ, -siːiː/) or Gramineae (/ɡrəˈmɪniaɪ/) is a large and nearly ubiquitous family of monocotyledonous flowering plants commonly known as grasses. It includes the cereal grasses, bamboos, the grasses of natural grassland and species cultivated in lawns and pasture. The latter are commonly referred to collectively as grass.

With around 780 genera and around 12,000 species,[4] the Poaceae is the fifth-largest plant family, following the Asteraceae, Orchidaceae, Fabaceae and Rubiaceae.[5]

The Poaceae are the most economically important plant family, providing staple foods from domesticated cereal crops such as maize, wheat, rice, barley, and millet as well as feed for meat-producing animals. They provide, through direct human consumption, just over one-half (51%) of all dietary energy; rice provides 20%,[6] wheat supplies 20%, maize (corn) 5.5%, and other grains 6%.[citation needed] Some members of the Poaceae are used as building materials (bamboo, thatch, and straw); others can provide a source of biofuel, primarily via the conversion of maize to ethanol.

Grasses have stems that are hollow except at the nodes and narrow alternate leaves borne in two ranks. The lower part of each leaf encloses the stem, forming a leaf-sheath. The leaf grows from the base of the blade, an adaptation allowing it to cope with frequent grazing.

Grasslands such as savannah and prairie where grasses are dominant are estimated to constitute 40.5% of the land area of the Earth, excluding Greenland and Antarctica.[7] Grasses are also an important part of the vegetation in many other habitats, including wetlands, forests and tundra.

Though they are commonly called "grasses", groups such as the seagrasses, rushes and sedges fall outside this family. The rushes and sedges are related to the Poaceae, being members of the order Poales, but the seagrasses are members of the order Alismatales. However, all of them belong to the monocot group of plants.